chore: mask MessageList with identifiers refactor
