+++ plugin-hrm-form/src/maskIdentifiers/index.ts
@@ -1,3 +1,4 @@
+/* eslint-disable import/no-unused-modules */
 /**
  * Copyright (C) 2021-2023 Technology Matters
  * This program is free software: you can redistribute it and/or modify
@@ -13,6 +14,70 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see https://www.gnu.org/licenses/.
  */
+import { Manager, DefaultTaskChannels, TaskChannelDefinition, MessagingCanvas, MessageList } from '@twilio/flex-ui';
 
-export * from './maskChannelStringsWithIdentifiers';
-export * from './maskManagerStringsWithIdentifiers';
+import { getInitializedCan, PermissionActions } from '../permissions';
+import { getAseloFeatureFlags, getTemplateStrings } from '../hrmConfig';
+
+// Mask identifiers in the channel templates
+export const maskChannelTemplatesWithIdentifiers = (channelType: TaskChannelDefinition) => {
+  const can = getInitializedCan();
+  const maskIdentifiers = !can(PermissionActions.VIEW_IDENTIFIERS);
+  if (!maskIdentifiers) return;
+
+  const {
+    IncomingTaskCanvas,
+    TaskListItem,
+    CallCanvas,
+    TaskCanvasHeader,
+    Supervisor,
+    TaskCard,
+    TaskInfoPanel,
+  } = channelType.templates;
+
+  IncomingTaskCanvas.firstLine = 'MaskIdentifiers';
+
+  CallCanvas.firstLine = 'MaskIdentifiers';
+
+  // Task list and panel when a call comes in
+  TaskListItem.firstLine = 'MaskIdentifiers';
+  TaskListItem.secondLine =
+    channelType === DefaultTaskChannels.Chat ? 'TaskLineWebChatAssignedMasked' : 'TaskLineChatAssignedMasked';
+
+  // Task panel during an active call
+  TaskCanvasHeader.title = 'MaskIdentifiers';
+  Supervisor.TaskCanvasHeader.title = 'MaskIdentifiers';
+
+  // Task Status in Agents page
+  if (!getAseloFeatureFlags().enable_teams_view_enhancements) TaskCard.firstLine = 'MaskIdentifiers';
+
+  TaskInfoPanel.content = 'TaskInfoPanelContentMasked';
+
+  Supervisor.TaskOverviewCanvas.firstLine = 'MaskIdentifiers';
+  Supervisor.TaskInfoPanel.content = 'TaskInfoPanelContentMasked';
+};
+
+export const maskManagerStringsWithIdentifiers = (newStrings: ReturnType<typeof getTemplateStrings>) => {
+  const can = getInitializedCan();
+  const maskIdentifiers = !can(PermissionActions.VIEW_IDENTIFIERS);
+  if (!maskIdentifiers) return;
+
+  const { strings } = Manager.getInstance();
+
+  // eslint-disable-next-line consistent-return
+  return Object.fromEntries(
+    Object.entries(strings).map(([key, value]) => [
+      key,
+      value.replace(/{{task.defaultFrom}}/g, newStrings.MaskIdentifiers),
+    ]),
+  );
+};
+
+export const maskMessageListWithIdentifiers = () => {
+  MessagingCanvas.defaultProps.memberDisplayOptions = {
+    theirDefaultName: 'XXXXXX',
+    theirFriendlyNameOverride: false,
+    yourFriendlyNameOverride: true,
+  };
+  MessageList.Content.remove('0');
+};
